no cast

+++ 14-binary_tree_balance.c
@@ -14,8 +14,8 @@ int binary_tree_balance(const binary_tree_t *tree)
 	if (tree == NULL)
 		return (0);
 
-	left_height = tree->left ? (int)binary_tree_height(tree->left) + 1 : 0;
-	right_height = tree->right ? (int)binary_tree_height(tree->right) + 1 : 0;
+	left_height = tree->left ? binary_tree_height(tree->left) + 1 : 0;
+	right_height = tree->right ? binary_tree_height(tree->right) + 1 : 0;
 
 	return (left_height - right_height);
 }
